Refactored spotify oauth flow to use router
