+++ server/index.js
@@ -1,18 +1,14 @@
-// Libraries & dependencies
+require('dotenv').config();
 const path = require('path');
 const cors = require('cors');
 const morgan = require('morgan');
 const express = require('express');
 const bodyParser = require('body-parser');
 const cookieParser = require('cookie-parser');
-const request = require('request');
-const querystring = require('querystring');
 const router = require('./router.js');
-require('dotenv').config();
 
 // Global vars
 const port = process.env.PORT || 3000;
-const stateKey = 'spotify_auth_state';
 
 // Express server + middleware + controllers
 const app = express()
@@ -22,90 +18,6 @@ const app = express()
   .use(bodyParser.json())
   .use(bodyParser.urlencoded({ extended: true }))
   .use(express.static(path.join(__dirname, '../client/dist/')))
-  .use('/', router)
-
-// Helper function for state serialization
-var generateRandomString = function(length) {
-  var text = '';
-  var possible = 'ABCDEFGHIJKLMNOPQRSTUVWXYZabcdefghijklmnopqrstuvwxyz0123456789';
-  for (var i = 0; i < length; i++) {
-    text += possible.charAt(Math.floor(Math.random() * possible.length));
-  }
-  return text;
-};
-
-// Login route to auth spotify account
-app.get('/login', (req, res) => {
-  let scope = 'user-read-private user-read-email playlist-modify-public playlist-modify-private';
-  let state = generateRandomString(16);
-  res.cookie(stateKey, state);
-
-  res.redirect('https://accounts.spotify.com/authorize?' +
-    querystring.stringify({
-      response_type: 'code',
-      client_id: process.env.CLIENT_ID,
-      scope: scope,
-      redirect_uri: process.env.REDIRECT_URI,
-      state: state
-    }));
-});
-
-// Redirect request to serve client token after getting auth code
-app.get('/callback', (req, res) => {
-  let code = req.query.code || null;
-  let storedState = req.cookies ? req.cookies[stateKey] : null;
-  res.clearCookie(stateKey);
-  let authOptions = {
-    json: true,
-    url: 'https://accounts.spotify.com/api/token',
-    headers: { 'Authorization': 'Basic ' + (new Buffer(process.env.CLIENT_ID + ':' + process.env.CLIENT_SECRET).toString('base64')) },
-    form: {
-      code: code,
-      redirect_uri: process.env.REDIRECT_URI,
-      grant_type: 'authorization_code'
-    },
-  };
-
-  request.post(authOptions, (err, resp, body) => {
-    if (!err && resp.statusCode === 200) {
-      let access_token = body.access_token;
-      let refresh_token = body.refresh_token;
-      let options = {
-        url: 'https://api.spotify.com/v1/me',
-        headers: { 'Authorization': 'Bearer ' + access_token },
-        json: true
-      };
-
-      // use the access token to access the Spotify Web API
-      // request.get(options, (err, resp, body) => console.log(body));
-      res.redirect('/#' + querystring.stringify({ access_token, refresh_token }));
-    } else {
-      res.redirect('/#' + querystring.stringify({ err: 'invalid_token' }));
-    }
-  });
-});
-
-// Refresh token endpoint that allows past users to get a fresh token w/o another oauth
-app.get('/refresh_token', (req, res) => {
-  let refresh_token = req.query.refresh_token;
-  let authOptions = {
-    json: true,
-    url: 'https://accounts.spotify.com/api/token',
-    headers: { 'Authorization': 'Basic ' + (new Buffer(process.env.CLIENT_ID + ':' + process.env.CLIENT_SECRET).toString('base64')) },
-    form: {
-      grant_type: 'refresh_token',
-      refresh_token: refresh_token
-    },
-  };
-
-  request.post(authOptions, (error, response, body) => {
-    if (!error && response.statusCode === 200) {
-      let access_token = body.access_token;
-      res.send({
-        'access_token': access_token
-      });
-    }
-  });
-});
+  .use(router)
 
 app.listen(port, () => console.log(`Listening on port: ${port}`));
